Move placeholder logic into shared class
