+++ grafast/dataplan-pg/src/pgLocker.ts
@@ -1,6 +1,6 @@
 import { isDev } from "grafast";
 
-import type { PgSelectStep, PgUnionAllStep } from "./index";
+import type { PgStmtBaseStep } from "./steps/pgStmt";
 
 export type PgLockableParameter =
   | "orderBy"
@@ -8,13 +8,11 @@ export type PgLockableParameter =
   | "last"
   | "offset"
   | "groupBy";
-export type PgLockCallback<
-  TStep extends PgSelectStep<any> | PgUnionAllStep<any, any>,
-> = (step: TStep) => void;
+export type PgLockCallback<TStep extends PgStmtBaseStep<any>> = (
+  step: TStep,
+) => void;
 
-export class PgLocker<
-  TStep extends PgSelectStep<any> | PgUnionAllStep<any, any>,
-> {
+export class PgLocker<TStep extends PgStmtBaseStep<any>> {
   /**
    * Determines if the PgSelectStep is "locked" - i.e. its
    * FROM,JOINs,WHERE,ORDER BY,LIMIT,OFFSET cannot be changed. Note this does
